backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
psycopg2-binary==2.9.9
alembic==1.13.2







